+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes
 
 ### Unreleased
 
+* 2022-12-14 - Feature: Built-in contact, help and maintenance pages, solves #150.
 * 2022-11-28 - Updated Moodle Plugin CI to latest upstream recommendations
 
 ### v4.0-r8